use float to preload scripts
