+++ packages/next/client/app-next.js
@@ -1,48 +1,6 @@
-const version = process.env.__NEXT_VERSION
+import { appBootstrap } from './app-bootstrap'
 
-window.next = {
-  version,
-  appDir: true,
-}
-
-function loadScriptsInSequence(scripts, hydrate) {
-  if (!scripts || !scripts.length) {
-    return hydrate()
-  }
-
-  return scripts
-    .reduce((promise, [src, props]) => {
-      return promise.then(() => {
-        return new Promise((resolve, reject) => {
-          const el = document.createElement('script')
-
-          if (props) {
-            for (const key in props) {
-              if (key !== 'children') {
-                el.setAttribute(key, props[key])
-              }
-            }
-          }
-
-          if (src) {
-            el.src = src
-            el.onload = resolve
-            el.onerror = reject
-          } else if (props) {
-            el.innerHTML = props.children
-            setTimeout(resolve)
-          }
-
-          document.head.appendChild(el)
-        })
-      })
-    }, Promise.resolve())
-    .then(() => {
-      hydrate()
-    })
-}
-
-loadScriptsInSequence(self.__next_scripts, () => {
+appBootstrap(() => {
   // Include app-router and layout-router in the main chunk
   import('next/dist/client/components/app-router.client.js')
   import('next/dist/client/components/layout-router.client.js')
