+++ packages/next/server/node-web-streams-helper.ts
@@ -306,7 +306,6 @@ export async function continueFromInitialStream(
     generateStaticHTML,
     getServerInsertedHTML,
     serverInsertedHTMLToHead,
-    polyfills,
   }: {
     dev?: boolean
     suffix?: string
@@ -314,7 +313,6 @@ export async function continueFromInitialStream(
     generateStaticHTML: boolean
     getServerInsertedHTML?: () => Promise<string>
     serverInsertedHTMLToHead: boolean
-    polyfills?: { src: string; integrity: string | undefined }[]
   }
 ): Promise<ReadableStream<Uint8Array>> {
   const closeTag = '</body></html>'
@@ -333,26 +331,13 @@ export async function continueFromInitialStream(
     dataStream ? createInlineDataStream(dataStream) : null,
     suffixUnclosed != null ? createSuffixStream(closeTag) : null,
     createHeadInjectionTransformStream(async () => {
-      // Inject polyfills for browsers that don't support modules. It has to be
-      // blocking here and can't be `defer` because other scripts have `async`.
-      const polyfillScripts = polyfills
-        ? polyfills
-            .map(
-              ({ src, integrity }) =>
-                `<script nomodule="">(self.__next_scripts=self.__next_scripts||[]).push(${JSON.stringify(
-                  [src, { integrity }]
-                )})</script>`
-            )
-            .join('')
-        : ''
-
       // TODO-APP: Insert server side html to end of head in app layout rendering, to avoid
       // hydration errors. Remove this once it's ready to be handled by react itself.
       const serverInsertedHTML =
         getServerInsertedHTML && serverInsertedHTMLToHead
           ? await getServerInsertedHTML()
           : ''
-      return polyfillScripts + serverInsertedHTML
+      return serverInsertedHTML
     }),
     dev ? createRootLayoutValidatorStream() : null,
   ].filter(nonNullable)
